Extract PlaceholderSplash
