+++ src/views/dev/index.ts
@@ -1,1 +1,2 @@
 export * from './BlankScreen'
+export * from './PlaceholderSplash'
